Moved files to another directories for restructuring
